@HowardHinnant 2020-02-08
Sorry for the delay in looking at this.

Your comment above about the entire db being read into memory on each iteration is incorrect.  `get_tzdb()` returns a reference to a function local static.  The initialization of this static is what reads the db into memory.  On subsequent calls, `get_tzdb()` simply returns a reference to that same static without re-initializing it.

I'm testing date/test/tz_test/validate.cpp with and without this patch and measuring virtually no difference in the execution times on average.  The variance between runs is much greater than the difference between the average of the runs.  I've tested with `AUTO_DOWNLOAD` set to both 0 and 1 to eliminate network effects.

While testing does appear to confirm your patch is correct (validate.cpp outputs the same answers), I'm hesitant to make this change as I'm not showing a clear benefit.