@jmc1283 2020-02-12
Yes, you are right about `get_tzdb()` returning a reference. Nevertheless, only calling it once eliminates some unnecessary function calls.

Maybe I wasn't clear in my initial comment, the function that I have really optimized is `get_local_time`. Since the rule lookup is an n time loop, the speed increase will vary by what time zone you are looking up. Because of that I'm not sure that the validate.cpp are the best tests for this PR, as my profiling of that file shows that only about 1/3 - 1/4 of the work that the validate.cpp is `get_local_time`. Despite this, on my system I saw a pretty significant 200ms difference in validate.cpp's execution time when comparing the master branch to my fork. 

This optimization shows its value when the loop in find_rule has many iterations by reducing unused work for zones with particularly slow lookups (like mine America/New_York).